+++ src/toisto/app.py
@@ -1,9 +1,9 @@
 """Main module for the application."""
 
 import readline  # pylint: disable=unused-import
-import requests
 
 from .command import practice, show_topics, show_progress
+from .metadata import latest_version
 from .persistence import load_topics, load_progress
 from .ui.cli import argument_parser
 from .ui.text import show_welcome
@@ -14,12 +14,8 @@ def main():
     args = argument_parser.parse_args()
     topics = load_topics(args.language, args.source_language, args.topic, args.topic_file)
     progress = load_progress(topics)
-    try:
-        print(requests.get("https://api.github.com/repos/fniessink/toisto/tags").json()[0]["name"])
-    except requests.ConnectionError:
-        pass
     if args.command == "practice":
-        show_welcome()
+        show_welcome(latest_version())
         practice(progress)
     elif args.command == "topics":
         show_topics(args.language, args.source_language, topics)